initial refactoring for harmony support
