+++ icepyx/__init__.py
@@ -1,5 +1,6 @@
 from _icepyx_version import version as __version__
-from icepyx.core.query import GenQuery, Query
+from icepyx.core.base_query import GenQuery, BaseQuery
+from icepyx.core.query import Query
 from icepyx.core.read import Read
 from icepyx.core.variables import Variables
 from icepyx.quest.quest import Quest
