+++ icepyx/core/is2ref.py
@@ -11,6 +11,13 @@ from icepyx.core.urls import COLLECTION_SEARCH_BASE_URL, EGI_BASE_URL
 
 # ICESat-2 specific reference functions
 
+def _get_concept_id(product, version):
+    collections = earthaccess.search_datasets(short_name=product,
+                                              version=version,
+                                              cloud_hosted=True)
+    if collections:
+        return collections[0].concept_id()
+
 
 def _validate_product(product):
     """
