+++ icepyx/quest/quest.py
@@ -1,4 +1,5 @@
-from icepyx.core.query import GenQuery, Query
+from icepyx.core.base_query import GenQuery
+from icepyx.core.query import Query
 from icepyx.quest.dataset_scripts.argo import Argo
