+++ requirements.txt
@@ -1,7 +1,7 @@
 backoff
 dask[dataframe]
 datashader
-earthaccess>=0.5.1
+earthaccess>=0.12.0
 fiona
 geopandas
 h5netcdf
@@ -15,3 +15,4 @@ requests
 s3fs
 shapely
 xarray
+deprecated
